vehicle/viss/0190-subscribe-notification-success.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>
[0190: receive the value continuously]
</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">
<link rel="stylesheet" type="text/css" href="./style-general.css">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-config.js"></script>
<script src="./vehicle-test-helper.js"></script>

<div id="test-title">
[0190: receive the value continuously]
</div>
<br>
<div id="test-assertion">
<b>Test assertion:</b><br>
When "subscribe" request has been successed,<br>
 client will receive a notificaton with the JSON data object continuously.<br>
</div>
<div id="scenario">
<br>
<b>Test Scenario:</b><br>
1. Execute 'subscribe' method specifing valid information.<br>
   (e.g. path = 'Signal.Drivetrain.Transmission.Speed')<br>
2. Confirm 'SubscribeSuccessResponse' is returned.<br>
3. Confirm 'SubscriptionNotification' is returned.<br>
4. Execute 'unsubscribe' method spedifing subscriptionId obtained at Step2.<br>
<br>
<b>Expected Result:</b><br>
- After 'subscribe' method succeeded, receives 'SubscriptionNotification' multiple times.<br>
</div>

<div id="result"></div>
<br>
<div id="log"></div>

<script>
// request a signal
var reqId = "reqid-0190";
var subId = null;
//var path = "Signal.Drivetrain.Transmission.Speed";
var path = SUBSCRIBE_STANDARD_PATH;

// phase
var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var NOTIFICATION_COUNT  = 3;

var notifCount = 0;
var phase = PH_START;

var t = async_test("Subscribe test: success 01");
var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

vehicle.onopen = function() {
  // process messages from the server
  vehicle.onmessage = function(event) {
    var msg = JSON.parse(event.data);
    addLogMessage("Response: json="+event.data);

    if (phase == PH_START) {
      // nothing to do
    } else if (phase == PH_SENT_SUBSCRIBE) {
      t.step_func(function() {
        // wait for subscribe response
        if (msg.action !== "subscribe") {
          addLogMessage("PH_SENT_SUBSCRIBE: Ignore messages except for Subscribe responses.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        var res = isSubscribeSuccessResponse(reqId, msg);
        assert_true(res, "True if subscribeSuccessResponse received.");
        if (res) {
          addLogMessage("subscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId
          phase = PH_SUBSCRIBED;
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          helper_terminate_failure("subscribeErrorResponse received. test failed.");
          phase = PH_END;
        }
      })();
    } else if (phase == PH_SUBSCRIBED) {
      t.step_func(function() {
        // wait subscriptionNotification
        if (msg.action !== "subscription") {
          addLogMessage("PH_SUBSCRIBED: Ignore messages except for Subscription notification.");
          addLogMessage("msg = " + JSON.stringify(msg));
          return;
        }
        var res = isSubscriptionNotificationResponse(subId, msg);
        assert_true(res, "True if subscriptionNotification received.");
        if (res){
          // after received 3 notif, send unsubscribe
          if (notifCount === NOTIFICATION_COUNT) {
            // do unsubscribe
            addLogMessage("<br>Send unsubscribe request");
            var unsub_request = '{"action":"unsubscribe"'
                               +' ,"subscriptionId":"'+subId+'"'
                               +' ,"requestId":"'+reqId+'"'
                               +'}';
            addLogMessage("Unsub request: json=" + unsub_request);
            vehicle.send( unsub_request );
            phase = PH_SENT_UNSUBSCRIBE;
          }
          addLogSuccess("subscriptionNotification received. value = " + msg.value);
          notifCount++;
        }
      })();
    } else if (phase == PH_SENT_UNSUBSCRIBE) {
      t.step_func(function() {
        // wait for unsubscribe response
        if (msg.action !== "unsubscribe") {
          addLogMessage("PH_SENT_UNSUBSCRIBE: Ignore messages except for Unsubscribe response.");
          addLogMessage("  msg = " + JSON.stringify(msg));
          return;
        }
        if (isUnsubscribeSuccessResponse(reqId, subId, msg)) {
          // unsubscibe success
          phase = PH_UNSUBSCRIBED;
          helper_terminate_normal("unsubscribe successed");
        } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
          // unsubscribe error
          // Test has been success already. Terminate anyway.
          helper_terminate_normal("unsubscribeErrorResponse received");
        }
      })();
    } else if (phase == PH_UNSUBSCRIBED) {
      // nothing to do
    }
  }

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {
      helper_terminate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME);

  addLogMessage("<br>Send subscribe request: path="+path);
  var sub_request = '{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+reqId+'"}';
  addLogMessage("Subscribe request: json="+sub_request);
  vehicle.send(sub_request);
  phase = PH_SENT_SUBSCRIBE;
}
</script>
</html>
